Move message types and functions to a new module #36
