+++ haskell-src-transformations.cabal
@@ -109,6 +109,7 @@ library
                     , HST.Frontend.Transformer
                     , HST.Frontend.Transformer.Messages
                     , HST.Options
+                    , HST.Util.Messages
                     , HST.Util.Predicates
                     , HST.Util.PrettyName
                     , HST.Util.Selectors
