+++ src/lib/HST/Feature/Optimization.hs
@@ -8,10 +8,10 @@ import           Polysemy                 ( Member, Members, Sem )
 import           HST.Effect.Fresh         ( Fresh )
 import           HST.Effect.PatternStack
   ( PatternStack, peekPattern, popPattern, pushPattern, runPatternStack )
-import           HST.Effect.Report
-  ( Report, Severity(Error), message, reportFatal )
+import           HST.Effect.Report        ( Report, reportFatal )
 import           HST.Environment.Renaming ( rename, subst )
 import qualified HST.Frontend.Syntax      as S
+import           HST.Util.Messages        ( Severity(Error), message )
 import           HST.Util.Selectors
   ( expFromUnguardedRhs, getAltConName, getPatConName, getPatVarName )
