+++ src/lib/HST/Frontend/GHC/To.hs
@@ -21,8 +21,7 @@ import qualified SrcLoc                  as GHC
 import qualified TcEvidence              as GHC
 import qualified TysWiredIn              as GHC
 
-import           HST.Effect.Report
-  ( Report, Severity(Internal), message, reportFatal )
+import           HST.Effect.Report       ( Report, reportFatal )
 import           HST.Frontend.GHC.Config
   ( DeclWrapper(Decl), GHC, LitWrapper(Lit, OverLit)
   , OriginalModuleHead(originalModuleName, originalModuleExports,
@@ -31,6 +30,7 @@ import           HST.Frontend.GHC.Config
   , TypeWrapper(SigType) )
 import qualified HST.Frontend.GHC.From   as FromGHC
 import qualified HST.Frontend.Syntax     as S
+import           HST.Util.Messages       ( Severity(Internal), message )
 
 -------------------------------------------------------------------------------
 -- Modules                                                                   --
