+++ src/lib/HST/Options.hs
@@ -21,9 +21,9 @@ import           Polysemy              ( Member, Sem )
 import           System.Console.GetOpt
   ( ArgDescr(NoArg, ReqArg), ArgOrder(Permute), OptDescr(Option), getOpt )
 
-import           HST.Effect.Report
-  ( Report, Severity(Error), message, report, reportFatal )
+import           HST.Effect.Report     ( Report, report, reportFatal )
 import qualified HST.Frontend.Syntax   as S
+import           HST.Util.Messages     ( Severity(Error), message )
 
 -- | A data type for all front ends that can be used for parsing the given input
 --   program in @haskell-src-transformations@.
